fixed: Toolbar and rolemanage complete

+++ client/src/pages/Article/AllArticle.jsx
@@ -3,7 +3,15 @@ import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import Toolbar from "../../features/dashboard/components/Toolbar";
 import ArticleCard from "../../features/dashboard/components/ArticleCard";
-import { getAllArticles, getAArticles, deleteArticle, addFavourite } from "../../redux/articleSlice";
+import {
+  getAllArticles,
+  getAArticles,
+  deleteArticle,
+  addFavourite,
+  setSearchVal,
+  setSortIndex,
+  setCategoryIndex,
+} from "../../redux/articleSlice";
 import { setPageTitle } from "../../features/common/headerSlice";
 // import { setIsLoading } from "../../redux/articleSlice";
 
@@ -18,6 +26,10 @@ function AllArticle() {
   const setHandleAddArticle = () => {
     navigate(`/newArticle/${0}`);
   };
+  const setShowArticle = (index) => {
+    dispatch(getAArticles(index));
+    navigate(`/showArticle/${index}`);
+  };
   const onFavouriteArticle = (index, from) => {
     dispatch(addFavourite({ index, from: user._id }));
   };
@@ -26,13 +38,26 @@ function AllArticle() {
     navigate(`/answerArticle/${index}`);
   };
   const setHandleDelete = (index) => {
-    if (window.confirm("Are you delete this Article?")) {
-      dispatch(deleteArticle(index));
-    }
+    if (user.roll !== "user") {
+      if (window.confirm("Are you delete this Article?")) {
+        dispatch(deleteArticle(index));
+      }
+    } else alert("You can't delete other articles, Now your role is user");
   };
   const setHandleEdit = (index) => {
-    dispatch(getAArticles(index));
-    navigate(`/newArticle/${index}`);
+    if (user.roll !== "user") {
+      dispatch(getAArticles(index));
+      navigate(`/newArticle/${index}`);
+    } else alert("You can't edit other articles, Now your role is user");
+  };
+  const setHandleSearchVal = (e) => {
+    dispatch(setSearchVal(e.target.value));
+  };
+  const setHnadleSortIndex = (e) => {
+    dispatch(setSortIndex(e.target.value));
+  };
+  const setHanldeCategoryIndex = (e) => {
+    dispatch(setCategoryIndex(e.target.value));
   };
   useEffect(() => {
     dispatch(getAllArticles());
@@ -40,8 +65,13 @@ function AllArticle() {
 
   return (
     <>
-      <Toolbar onAddClick={setHandleAddArticle} />
-      {value.article.length &&
+      <Toolbar
+        setSortIndex={setHnadleSortIndex}
+        setCategoryIndex={setHanldeCategoryIndex}
+        setSearchval={setHandleSearchVal}
+        onAddClick={setHandleAddArticle}
+      />
+      {value.article.length > 0 ? (
         value.article.map((v, i) => {
           return (
             <div key={i}>
@@ -55,6 +85,9 @@ function AllArticle() {
                 onFavouriteClick={() => {
                   onFavouriteArticle(v._id, v.from._id);
                 }}
+                setShowArticle={() => {
+                  setShowArticle(v._id);
+                }}
                 onDeleteArticle={() => {
                   setHandleDelete(v._id);
                 }}
@@ -67,7 +100,10 @@ function AllArticle() {
               />
             </div>
           );
-        })}
+        })
+      ) : (
+        <div className="text-[50px] text-red-500">No Data</div>
+      )}
     </>
   );
 }
